+++ versions/v20/projects/sdcorejs-angular/components/table/index.ts
@@ -1,6 +1,7 @@
 export * from './src/directives/sd-table-cell-def.directive';
 export * from './src/directives/sd-table-row-mobile-def.directive';
 export * from './src/directives/sd-table-command-header-def.directive';
+export * from './src/directives/sd-table-quick-search-right-def.directive';
 export * from './src/directives/sd-table-expand-def.directive';
 export * from './src/directives/sd-table-filter-def.directive';
 export * from './src/directives/sd-table-footer-def.directive';